+++ demo/symfony8/config/bundles.php
@@ -1,10 +1,12 @@
 <?php
 
 declare(strict_types=1);
+
 use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
 use Doctrine\Bundle\FixturesBundle\DoctrineFixturesBundle;
 use Doctrine\Bundle\MigrationsBundle\DoctrineMigrationsBundle;
 use Nowo\AuthKitBundle\NowoAuthKitBundle;
+use Nowo\FormKitBundle\NowoFormKitBundle;
 use Nowo\PasswordToggleBundle\NowoPasswordToggleBundle;
 use Nowo\TwigInspectorBundle\NowoTwigInspectorBundle;
 use Symfony\Bundle\DebugBundle\DebugBundle;
@@ -13,8 +15,10 @@ use Symfony\Bundle\SecurityBundle\SecurityBundle;
 use Symfony\Bundle\TwigBundle\TwigBundle;
 use Symfony\Bundle\WebProfilerBundle\WebProfilerBundle;
 use Symfony\UX\Icons\UXIconsBundle;
+use Twig\Extra\TwigExtraBundle\TwigExtraBundle;
 
 return [
+    NowoFormKitBundle::class        => ['all' => true],
     FrameworkBundle::class          => ['all' => true],
     DoctrineBundle::class           => ['all' => true],
     DoctrineMigrationsBundle::class => ['all' => true],
@@ -27,4 +31,5 @@ return [
     WebProfilerBundle::class        => ['dev' => true, 'test' => true],
     DebugBundle::class              => ['dev' => true],
     NowoPasswordToggleBundle::class => ['all' => true],
+    TwigExtraBundle::class          => ['all' => true],
 ];