+++ src/Form/ResetPasswordFormType.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace Nowo\AuthKitBundle\Form;
 
 use Nowo\AuthKitBundle\NowoAuthKitBundle;
+use Nowo\FormKitBundle\Attribute\FormKitConfig;
+use Nowo\FormKitBundle\Form\FormOptionsTrait;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
@@ -12,8 +14,11 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
 /**
  * Form to set a new password after following a reset link.
  */
+#[FormKitConfig('auth_kit')]
 final class ResetPasswordFormType extends AbstractType
 {
+    use FormOptionsTrait;
+
     public function __construct(
         private readonly PasswordRepeatedFieldBuilder $passwordRepeatedFieldBuilder,
     ) {